backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.5
python-dotenv==1.0.0
aiosmtplib==3.0.1
email-validator==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
